Fixed bug: Overflowing 'int' in Controller 'nodes' map. Changed to 'long long'.

+++ Controller.h
@@ -26,7 +26,7 @@ class Controller {
     private:
         std::mutex mapMutex; // Mutex to lock access to map
         // Records storage node info (i.e. port, space, lastbeat, state)
-        std::map<int, std::map<std::string, int>> nodes;
+        std::map<int, std::map<std::string, long long>> nodes;
         // Records the files saved at storage nodes
         std::map<int,std::vector<std::string>> inventory;
         int newSocket, addressLen, opt;
